requirements.txt
tornado
mcp
openai
pathspec
rich
anthropic
python-dotenv
pytest
